+++ NEWS.md
@@ -1,5 +1,10 @@
 # rsconnect 0.8.30 (development version)
 
+* Deployment records no longer contain the time the app was deployed (`when`)
+  or when it's metadata was last synched (`lastSyncTime`) as these variables
+  are not very useful, and they lead to uninteresting diffs if you have 
+  committed the deployment records to git (#770).
+
 * `deployApp()` will now detect if you're attempting to publish to an app
   that has been deleted and will prompt you to create a new app (#226).
 
@@ -38,8 +43,6 @@
 
 * `deployApp()` now advertises which startup scripts are run at the normal
   `logLevel`, and it evaluates each script in its own environment (#542).
-
-* `deployments()` now formats `when` and `lastSyncTime` as date-times (#714).
 
 * `deployApp()` now derives `appName` from `appDir` and `appPrimaryDoc`, 
   never using the title (#538). It now only simplifies the path if you are